+++ src/classes_units.py
@@ -6,6 +6,7 @@ from settings import *
 from functions_math import *
 from functions_player import *
 from classes_vehicles import *
+from classes_planes import *
 from classes_turrets import *
 
 
@@ -80,6 +81,9 @@ class Unit:
         self.HP -= power
         if self.HP <= 0:
             self.is_alive = False
+
+
+# ======================================================================
 
 
 class Land_unit(Unit):
@@ -164,3 +168,34 @@ class Spider_tank(Land_unit):
         # draw unit indicator
         pygame.draw.line(win, WHITE, [coord_on_screen[0] - 3, coord_on_screen[1]], [coord_on_screen[0] + 3, coord_on_screen[1]], 1) # -
         pygame.draw.line(win, WHITE, [coord_on_screen[0], coord_on_screen[1] - 3], [coord_on_screen[0], coord_on_screen[1] + 3], 1) # |
+
+
+# ======================================================================
+
+
+class Air_unit(Unit):
+    pass
+
+
+class Fighter(Air_unit):
+    Vehicle_class = Plane
+    Turret_class = Minigun
+
+
+    def draw(self, win, offset_x, offset_y, scale):
+    # draw the spider tank on the screen
+        Air_unit.draw(self, win, offset_x, offset_y, scale)
+
+        coord_on_screen = world2screen(self.coord, offset_x, offset_y, scale)
+
+        # draw level indicator
+        pygame.draw.line(win, BLACK, coord_on_screen, [coord_on_screen[0], coord_on_screen[1] + 9], 7)
+        pygame.draw.line(win, WHITE, [coord_on_screen[0] - 2, coord_on_screen[1]], [coord_on_screen[0] - 2, coord_on_screen[1] + 8], 2)
+        pygame.draw.line(win, WHITE, [coord_on_screen[0] + 1, coord_on_screen[1]], [coord_on_screen[0] + 1, coord_on_screen[1] + 8], 2)
+
+        # draw team circle
+        pygame.draw.circle(win, player_color(self.player_id), coord_on_screen, 7, 0)
+
+        # draw unit indicator
+        pygame.draw.line(win, WHITE, [coord_on_screen[0] - 3, coord_on_screen[1] + 3], [coord_on_screen[0], coord_on_screen[1] - 3], 1) # /
+        pygame.draw.line(win, WHITE, [coord_on_screen[0] + 3, coord_on_screen[1] + 3], [coord_on_screen[0], coord_on_screen[1] - 3], 1) # \